change file to multiple files
